Elimination heuristic
The heuristic is now computed as follows: If a color can be eliminated
in one move, do that since that is an optimal move. Else, do the usual
"color-blind move".

Every color has to be eliminated at some point, so this heuristic is at
least as good as the previous one. In practice, a collection of 1000
puzzles à 14x14 with 6 colors is solved roughly twice as fast.