Eliminate unnecessary copying of state
Unfortunately it's impossible to directly move out of a priority_queue,
so we have to go via the underlying vector instead. Fortunately that's
not very hard though, since priority_queue is just a thin wrapper using
standard algorithms like push_heap and pop_heap which we can also use
directly.